moved google related logic to a submodule
