+++ oxidauth-kernel/src/auth/authenticate_or_register.rs
@@ -77,9 +77,11 @@ pub struct OAuth2AuthenticatePathParams {
     pub scope: String,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-pub struct OAuth2AuthenticateResponse {
-    pub profile: JsonValue,
+#[derive(Debug, Clone, Deserialize)]
+pub struct OAuth2Profile {
+    pub email: String,
+    pub given_name: String,
+    pub family_name: String,
 }
 
 #[derive(Debug)]
@@ -94,44 +96,5 @@ impl fmt::Display for ParseOAuth2AuthenticateUrlError {
         match self {
             Unknown(value) => write!(f, "unable to handle oauth authenticate response: {}", value),
         }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GoogleExchangeTokenReq {
-    pub code: String,
-    pub client_id: String,
-    pub client_secret: String,
-    pub redirect_uri: String,
-    pub grant_type: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GoogleExchangeTokenRes {
-    pub access_token: String,
-    pub expires_in: u32,
-    pub scope: String,
-    pub token_type: String,
-    pub id_token: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GoogleProfile {
-    pub name: String,
-    pub given_name: String,
-    pub family_name: String,
-    pub picture: String,
-    pub id: String,
-    pub email: String,
-    pub verified_email: bool,
-}
-
-impl TryFrom<JsonValue> for GoogleProfile {
-    type Error = BoxedError;
-
-    fn try_from(value: JsonValue) -> Result<Self, Self::Error> {
-        let profile = serde_json::from_value(value.inner_value())?;
-
-        Ok(profile)
     }
 }
