+++ oxidauth-usecases/src/auth/mod.rs
@@ -1,4 +1,5 @@
 pub mod authenticate;
 pub mod authenticate_or_register;
+pub mod google;
 pub mod register;
 pub mod strategies;
